minor changes on optimization and code-style
