+++ example/node_example/example_with_creation_from_setters.js
@@ -55,7 +55,7 @@ const sendACaptureRequestExample = async () => {
 
   console.log(`BizStep of the object event: ${objectEvent.getBizStep()}`);
   console.log(`Action of the object event: ${objectEvent.getAction()}`);
-  console.log(`objectEvent: `);
+  console.log('objectEvent: ');
   console.log(objectEvent.toObject());
   console.log(`epcisDocument (toString): ${epcisDocument.toString()}`);
