+++ src/hash_generator/dlNormalizer.js
@@ -37,11 +37,7 @@ const getCheckDigit = (input, throwError) => {
     let i = 1;
     array.forEach((number) => {
       const n = parseInt(number, 10);
-      if (i % 2 === 0) {
-        total += n;
-      } else {
-        total += n * 3;
-      }
+      total += (i % 2 === 0) ? n : 3 * n;
       i += 1;
     });
     return Math.ceil(total / 10) * 10 - total;
@@ -73,6 +69,9 @@ export const addCheckDigitAndZeroPad = (raw, throwError, length = 14) => {
 
 /**
  * Converts any standard URI conveying a GS1 Key in Canonical GS1 DL URI.
+ *
+ * For example, it transforms 'urn:epc:id:sgtin:0614141.011111.987' to 'https://id.gs1.org/01/00614141111114/21/987'
+ *
  * @param originalUri - any URI to be used in EPCIS events that convey a GS1 key, i.e. EPC URIs, EPC
  * Class URIs, EPC ID Pattern URIs, or GS1 Digital Link URIs.
  * @param {boolean} throwError - if set to true, it will throw an error if the event misses some
@@ -90,6 +89,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
   const partition = uri.indexOf('.');
 
   // EPC URIs
+  // if the input is a valid sgtin urn
   if (uri.match('^urn:epc:id:sgtin:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20}$')) {
     const gs1CompanyPrefix = uri.substring(17, partition);
     const itemRef = uri.substring(partition + 1, partition + 1 + (13 - gs1CompanyPrefix.length));
@@ -98,6 +98,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/01/${addCheckDigitAndZeroPad(rawGTIN, throwError)}/21/${webUriPercentEncoder(serial)}`);
   }
 
+  // if the input is a valid sscc urn
   if (uri.match('^urn:epc:id:sscc:(([0-9]{6}.[0-9]{11})|([0-9]{7}.[0-9]{10})|([0-9]{8}.[0-9]{9})|([0-9]{9}.[0-9]{8})|([0-9]{10}.[0-9]{7})|([0-9]{11}.[0-9]{6})|([0-9]{12}.[0-9]{5}))$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const serialRef = uri.substring(partition + 1);
@@ -105,6 +106,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/00/${addCheckDigitAndZeroPad(rawSSCC, throwError, 18)}`);
   }
 
+  // if the input is a valid sgln urn
   if (uri.match('^urn:epc:id:sgln:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}.)).(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const locationRef = uri.substring(partition + 1,
@@ -115,6 +117,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return `https://id.gs1.org/414/${addCheckDigitAndZeroPad(rawGLN, throwError, 13)}/254/${webUriPercentEncoder(extension)}`;
   }
 
+  // if the input is a valid grai urn
   if (uri.match('^urn:epc:id:grai:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}.)).(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,16}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const assetRef = uri.substring(partition + 1, (partition + 1 + (12 - gs1CompanyPrefix.length)));
@@ -123,12 +126,14 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return `https://id.gs1.org/8003/${addCheckDigitAndZeroPad(rawGRAI, throwError)}${webUriPercentEncoder(serial)}`;
   }
 
+  // if the input is a valid giai urn
   if (uri.match('^urn:epc:id:giai:(([0-9]{6}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,24})|([0-9]{7}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,23})|([0-9]{8}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,22})|([0-9]{9}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,21})|([0-9]{10}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20})|([0-9]{11}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,19})|([0-9]{12}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,18}))$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const assetRef = uri.substring(partition + 1);
     return (`https://id.gs1.org/8004/${gs1CompanyPrefix}${webUriPercentEncoder(assetRef)}`);
   }
 
+  // if the input is a valid gsrn urn
   if (uri.match('^urn:epc:id:gsrn:(([0-9]{6}.[0-9]{11})|([0-9]{7}.[0-9]{10})|([0-9]{8}.[0-9]{9})|([0-9]{9}.[0-9]{8})|([0-9]{10}.[0-9]{7})|([0-9]{11}.[0-9]{6})|([0-9]{12}.[0-9]{5}))$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const serviceRef = uri.substring(partition + 1);
@@ -136,6 +141,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8018/${addCheckDigitAndZeroPad(rawGSRN, throwError, 18)}`);
   }
 
+  // if the input is a valid gsrnp urn
   if (uri.match('^urn:epc:id:gsrnp:(([0-9]{6}.[0-9]{11})|([0-9]{7}.[0-9]{10})|([0-9]{8}.[0-9]{9})|([0-9]{9}.[0-9]{8})|([0-9]{10}.[0-9]{7})|([0-9]{11}.[0-9]{6})|([0-9]{12}.[0-9]{5}))$')) {
     const gs1CompanyPrefix = uri.substring(17, partition);
     const serviceRef = uri.substring(partition + 1);
@@ -143,6 +149,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8017/${addCheckDigitAndZeroPad(rawGSRNP, throwError, 18)}`);
   }
 
+  // if the input is a valid gdti urn
   if (uri.match('^urn:epc:id:gdti:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}..))(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const documentType = uri.substring(partition + 1,
@@ -152,6 +159,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return `https://id.gs1.org/253/${addCheckDigitAndZeroPad(rawGDTI, throwError, 13)}${webUriPercentEncoder(serial)}`;
   }
 
+  // if the input is a valid cpi urn
   if (uri.match('^urn:epc:id:cpi:(([0-9]{6}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,24})|([0-9]{7}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,23})|([0-9]{8}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,22})|([0-9]{9}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,21})|([0-9]{10}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,20})|([0-9]{11}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,19})|([0-9]{12}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,18})).[0-9]{1,12}$')) {
     const gs1CompanyPrefix = uri.substring(15, partition);
     const separator = uri.lastIndexOf('.');
@@ -161,6 +169,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8010/${webUriPercentEncoder(rawCPI)}/8011/${serial}`);
   }
 
+  // if the input is a valid sgcn urn
   if (uri.match('^urn:epc:id:sgcn:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}.)).[0-9]{1,12}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const couponRef = uri.substring(partition + 1,
@@ -170,12 +179,14 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/255/${addCheckDigitAndZeroPad(rawSGCN, throwError, 13)}${serial}`);
   }
 
+  // if the input is a valid ginc urn
   if (uri.match('^urn:epc:id:ginc:(([0-9]{6}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,24})|([0-9]{7}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,23})|([0-9]{8}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,22})|([0-9]{9}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,21})|([0-9]{10}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20})|([0-9]{11}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,19})|([0-9]{12}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,18}))$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const consignmentRef = uri.substring(partition + 1);
     return (`https://id.gs1.org/401/${gs1CompanyPrefix}${webUriPercentEncoder(consignmentRef)}`);
   }
 
+  // if the input is a valid gsin urn
   if (uri.match('^urn:epc:id:gsin:(([0-9]{6}.[0-9]{10}$)|([0-9]{7}.[0-9]{9}$)|([0-9]{8}.[0-9]{8}$)|([0-9]{9}.[0-9]{7}$)|([0-9]{10}.[0-9]{6}$)|([0-9]{11}.[0-9]{5}$)|([0-9]{12}.[0-9]{4}$))')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const shipperRef = uri.substring(partition + 1);
@@ -183,6 +194,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/402/${addCheckDigitAndZeroPad(rawGSIN, throwError, 17)}`);
   }
 
+  // if the input is a valid itip urn
   if (uri.match('^urn:epc:id:itip:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).[0-9]{2}.[0-9]{2}.(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -194,6 +206,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     }${piece}${total}/21/${webUriPercentEncoder(serial)}`);
   }
 
+  // if the input is a valid upui urn
   if (uri.match('^urn:epc:id:upui:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,28}$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -202,6 +215,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/01/${addCheckDigitAndZeroPad(rawGTIN, throwError)}/235/${webUriPercentEncoder(serial)}`);
   }
 
+  // if the input is a valid pgln urn
   if (uri.match('^urn:epc:id:pgln:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}.))$')) {
     const gs1CompanyPrefix = uri.substring(16, partition);
     const partyRef = uri.substring(partition + 1, (partition + 1 + (12 - gs1CompanyPrefix.length)));
@@ -209,6 +223,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/417/${addCheckDigitAndZeroPad(rawGLN, throwError, 13)}`);
   }
 
+  // if the input is a valid lgtin urn
   if (uri.match('^urn:epc:class:lgtin:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).(%2[125-9A-Fa-f]|%3[0-9A-Fa-f]|%4[1-9A-Fa-f]|%5[0-9AaFf]|%6[1-9A-Fa-f]|%7[0-9Aa]|[!\')(*+,.0-9:;=A-Za-z_-]){1,20}$')) {
     const gs1CompanyPrefix = uri.substring(20, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -217,6 +232,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/01/${addCheckDigitAndZeroPad(rawGTIN, throwError)}/10/${webUriPercentEncoder(lot)}`);
   }
 
+  // if the input is a valid sgtin (idpat) urn
   if (uri.match('^urn:epc:idpat:sgtin:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).\*$')) {
     const gs1CompanyPrefix = uri.substring(20, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -224,6 +240,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/01/${addCheckDigitAndZeroPad(rawGTIN, throwError)}`);
   }
 
+  // if the input is a valid grai (idpat) urn
   if (uri.match('^urn:epc:idpat:grai:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}..)).\*$')) {
     const gs1CompanyPrefix = uri.substring(19, partition);
     const assetRef = uri.substring(partition + 1, (partition + 1 + (12 - gs1CompanyPrefix.length)));
@@ -231,6 +248,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8003/${addCheckDigitAndZeroPad(rawGRAI, throwError)}`);
   }
 
+  // if the input is a valid gdti (idpat) urn
   if (uri.match('^urn:epc:idpat:gdti:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}..)).\*$')) {
     const gs1CompanyPrefix = uri.substring(19, partition);
     const documentType = uri.substring(partition + 1, (
@@ -239,6 +257,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/253/${addCheckDigitAndZeroPad(rawGDTI, throwError, 13)}`);
   }
 
+  // if the input is a valid sgcn (idpat) urn
   if (uri.match('^urn:epc:idpat:sgcn:(([0-9]{6}.[0-9]{6})|([0-9]{7}.[0-9]{5})|([0-9]{8}.[0-9]{4})|([0-9]{9}.[0-9]{3})|([0-9]{10}.[0-9]{2})|([0-9]{11}.[0-9]{1})|([0-9]{12}..)).\*$')) {
     const gs1CompanyPrefix = uri.substring(19, partition);
     const couponRef = uri.substring(
@@ -248,6 +267,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/255/${addCheckDigitAndZeroPad(rawSGCN, throwError, 13)}`);
   }
 
+  // if the input is a valid cpi (idpat) urn
   if (uri.match('^urn:epc:idpat:cpi:(([0-9]{6}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,24})|([0-9]{7}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,23})|([0-9]{8}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,22})|([0-9]{9}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,21})|([0-9]{10}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,20})|([0-9]{11}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,19})|([0-9]{12}.(%2[3dfDF]|%3[0-9]|%4[1-9A-Fa-f]|%5[0-9Aa]|[0-9A-Z-]){1,18})).\*$')) {
     const gs1CompanyPrefix = uri.substring(18, partition);
     const separator = uri.lastIndexOf('.');
@@ -256,6 +276,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8010/${webUriPercentEncoder(rawCPI)}`);
   }
 
+  // if the input is a valid itip (idpat) urn
   if (uri.match('^urn:epc:idpat:itip:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).[0-9]{2}.[0-9]{2}.\*$')) {
     const gs1CompanyPrefix = uri.substring(19, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -265,6 +286,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/8006/${addCheckDigitAndZeroPad(rawGTIN, throwError)}${piece}${total}`);
   }
 
+  // if the input is a valid upui (idpat) urn
   if (uri.match('^urn:epc:idpat:upui:(([0-9]{6}.[0-9]{7})|([0-9]{7}.[0-9]{6})|([0-9]{8}.[0-9]{5})|([0-9]{9}.[0-9]{4})|([0-9]{10}.[0-9]{3})|([0-9]{11}.[0-9]{2})|([0-9]{12}.[0-9]{1})).\*$')) {
     const gs1CompanyPrefix = uri.substring(19, partition);
     const itemRef = uri.substring(partition + 1, (partition + 1 + (13 - gs1CompanyPrefix.length)));
@@ -272,6 +294,7 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     return (`https://id.gs1.org/01/${addCheckDigitAndZeroPad(rawGTIN, throwError)}`);
   }
 
+  // if the input isn't a gs1 digital link that contains a key that we need to change here
   if (!uri.match('^https?:(\/\/((([^\/?#]*)@)?([^\/?#:]*)(:([^\/?#]*))?))?((([^?#]*)(\/(01|gtin|8006|itip|8010|cpid|414|gln|417|party|8017|gsrnp|8018|gsrn|255|gcn|00|sscc|253|gdti|401|ginc|402|gsin|8003|grai|8004|giai)\/)([0-9]{4}[^\/]+)?[\/]?(([^?\n]*))?(#([^]*))?)|(\/[A-Za-z_-]{10}$))')) {
     return uri;
   }
@@ -280,63 +303,51 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     uri = uri.substring(0, uri.indexOf('?'));
   }
 
-  uri = uri.split('/gtin/').join('/01/')
-    .split('/itip/').join('/8006/')
-    .split('/cpid/')
-    .join('/8010/')
-    .split('/gln/')
-    .join('/414/')
-    .split('/party/')
-    .join('/417/')
-    .split('/gsrnp/')
-    .join('/8017/')
-    .split('/gsrn/')
-    .join('/8018/')
-    .split('/gcn/')
-    .join('/255/')
-    .split('/sscc/')
-    .join('/00/')
-    .split('/gdti/')
-    .join('/253/')
-    .split('/ginc/')
-    .join('/401/')
-    .split('/gsin/')
-    .join('/402/')
-    .split('/grai/')
-    .join('/8003/')
-    .split('/giai/')
-    .join('/8004/')
-    .split('/cpv/')
-    .join('/22/')
-    .split('/lot/')
-    .join('/10/')
-    .split('/ser/')
-    .join('/21/');
+  // Replace non-numeric codes by numeric codes
+  uri = uri.replace('/gtin/', '/01/')
+    .replace('/itip/', '/8006/')
+    .replace('/cpid/', '/8010/')
+    .replace('/gln/', '/414/')
+    .replace('/party/', '/417/')
+    .replace('/gsrnp/', '/8017/')
+    .replace('/gsrn/', '/8018/')
+    .replace('/gcn/', '/255/')
+    .replace('/sscc/', '/00/')
+    .replace('/gdti/', '/253/')
+    .replace('/ginc/', '/401/')
+    .replace('/gsin/', '/402/')
+    .replace('/grai/', '/8003/')
+    .replace('/giai/', '/8004/')
+    .replace('/cpv/', '/22/')
+    .replace('/lot/', '/10/')
+    .replace('/ser/', '/21/');
 
   if (!uri.match('^https:\/\/id.gs1.org\/(01|8006|8010|414|417|8017|8018|255|00|253|401|402|8003|8004)\/([0-9]{4}[^\/]+)(\/[^\/]+\/[^\/]+)?[\/]?(([^?\n]*))?(#([^\n]*))?|(\/[A-Za-z_-]{10}$)')) {
-    if (uri.indexOf('/00/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/00/'))}`;
-    else if (uri.indexOf('/01/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/01/'))}`;
-    else if (uri.indexOf('/253/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/253/'))}`;
-    else if (uri.indexOf('/255/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/255/'))}`;
-    else if (uri.indexOf('/401/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/401/'))}`;
-    else if (uri.indexOf('/402/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/402/'))}`;
-    else if (uri.indexOf('/414/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/414/'))}`;
-    else if (uri.indexOf('/417/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/417/'))}`;
-    else if (uri.indexOf('/8003/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8003/'))}`;
-    else if (uri.indexOf('/8004/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8004/'))}`;
-    else if (uri.indexOf('/8006/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8006/'))}`;
-    else if (uri.indexOf('/8010/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8010/'))}`;
-    else if (uri.indexOf('/8017/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8017/'))}`;
-    else if (uri.indexOf('/8018/') !== -1) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8018/'))}`;
+    if (uri.includes('/00/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/00/'))}`;
+    else if (uri.includes('/01/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/01/'))}`;
+    else if (uri.includes('/253/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/253/'))}`;
+    else if (uri.includes('/255/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/255/'))}`;
+    else if (uri.includes('/401/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/401/'))}`;
+    else if (uri.includes('/402/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/402/'))}`;
+    else if (uri.includes('/414/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/414/'))}`;
+    else if (uri.includes('/417/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/417/'))}`;
+    else if (uri.includes('/8003/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8003/'))}`;
+    else if (uri.includes('/8004/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8004/'))}`;
+    else if (uri.includes('/8006/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8006/'))}`;
+    else if (uri.includes('/8010/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8010/'))}`;
+    else if (uri.includes('/8017/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8017/'))}`;
+    else if (uri.includes('/8018/')) uri = `https://id.gs1.org${uri.substring(uri.indexOf('/8018/'))}`;
   }
 
+  // Zero pad GTIN
   if (!uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{14}')) {
-    if (uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{13}')) {
-      uri = uri.replace('/01/', '/01/0');
-    } else if (uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{12}')) {
-      uri = uri.replace('/01/', '/01/00');
-    } else if (uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{8}')) {
-      uri = uri.replace('/01/', '/01/000000');
+    if (
+      uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{13}') ||
+      uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{12}') ||
+      uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{8}')) {
+      while (!uri.match('^https:\/\/id.gs1.org\/01\/[0-9]{14}')) {
+        uri = uri.replace('/01/', '/01/0');
+      }
     }
   }
 
@@ -368,6 +379,8 @@ export const dlNormalizer = (originalUri, throwError = true) => {
     uri = uri.substring(0, uri.indexOf('/10/')) + y.substring(y.indexOf('/'));
   }
 
+  // Checks if the uri is now a valid one or still not. If it is still not a valid one,
+  // it returns the original one. Otherwise, it returns the updated one.
   if (
     uri.match('https:\/\/id.gs1.org\/00\/([0-9]{18})$') ||
     uri.match('https:\/\/id.gs1.org\/01\/([0-9]{14})\/21\/([\x22\x27\x2D\x2E\x30-\x39\x3B-\x3F\x41-\x5A\x5F\x61-\x7A]{0,20})$') ||
