+++ test/data/hashing/event-hash-reference-document3.json
